Usability enhancements, better default artwork
- Bring in animate.css, add some simple animations
- Remove images with unknown origins and copyright statuses
- Add new ribbon images and add retina versions
- Add focus to current patron's point earnings
- Add default avatars and new front page image (thanks @tinyspeck)
- Fix minor feed/status issues
- Fix avatar drop-downs
- Move CSS so that customized CSS loads with highest precedence
- Add support for JPG banners

+++ SRP_DAL/Programs.cs
@@ -843,9 +843,23 @@ namespace GRA.SRP.DAL
 
                 iReturn = SqlHelper.ExecuteNonQuery(conn, CommandType.StoredProcedure, "app_Programs_Delete", arrParams);
                 var fileName = (HttpContext.Current.Server.MapPath("~/Images/Banners/") + "\\" + PID.ToString() + ".png");
-                File.Delete(fileName);
+                if(File.Exists(fileName)) {
+                    File.Delete(fileName);
+                }
+                fileName = (HttpContext.Current.Server.MapPath("~/Images/Banners/") + "\\" + PID.ToString() + "@2x.png");
+                if(File.Exists(fileName)) {
+                    File.Delete(fileName);
+                }
+                fileName = (HttpContext.Current.Server.MapPath("~/Images/Banners/") + "\\" + PID.ToString() + ".jpg");
+                if(File.Exists(fileName)) {
+                    File.Delete(fileName);
+                }
+                fileName = (HttpContext.Current.Server.MapPath("~/Images/Banners/") + "\\" + PID.ToString() + "@2x.jpg");
+                if(File.Exists(fileName)) {
+                    File.Delete(fileName);
+                }
 
-                 fileName = (HttpContext.Current.Server.MapPath("~/css/program/") + "\\" + PID.ToString() + ".css");
+                fileName = (HttpContext.Current.Server.MapPath("~/css/program/") + "\\" + PID.ToString() + ".css");
                 File.Delete(fileName);
                  fileName = (HttpContext.Current.Server.MapPath("~/resources/") + "\\program." + PID.ToString() + "en-US.txt");
                 File.Delete(fileName);
